move allowed mana pledge
