+++ plugins/webapi/mana/allowedmanapledge.go
@@ -1,4 +1,4 @@
-package value
+package mana
 
 import (
 	"net/http"
@@ -9,7 +9,7 @@ import (
 
 	"github.com/iotaledger/goshimmer/packages/mana"
 	manaPlugin "github.com/iotaledger/goshimmer/plugins/messagelayer"
-	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels/value"
+	"github.com/iotaledger/goshimmer/plugins/webapi/jsonmodels"
 )
 
 // Handler handles the request.
@@ -20,7 +20,7 @@ func allowedManaPledgeHandler(c echo.Context) error {
 		accessNodes = append(accessNodes, base58.Encode(element.(identity.ID).Bytes()))
 	})
 	if len(accessNodes) == 0 {
-		return c.JSON(http.StatusNotFound, value.AllowedManaPledgeResponse{Error: "No access mana pledge IDs are accepted"})
+		return c.JSON(http.StatusNotFound, jsonmodels.AllowedManaPledgeResponse{Error: "No access mana pledge IDs are accepted"})
 	}
 
 	consensus := manaPlugin.GetAllowedPledgeNodes(mana.ConsensusMana)
@@ -29,15 +29,15 @@ func allowedManaPledgeHandler(c echo.Context) error {
 		consensusNodes = append(consensusNodes, base58.Encode(element.(identity.ID).Bytes()))
 	})
 	if len(consensusNodes) == 0 {
-		return c.JSON(http.StatusNotFound, value.AllowedManaPledgeResponse{Error: "No consensus mana pledge IDs are accepted"})
+		return c.JSON(http.StatusNotFound, jsonmodels.AllowedManaPledgeResponse{Error: "No consensus mana pledge IDs are accepted"})
 	}
 
-	return c.JSON(http.StatusOK, value.AllowedManaPledgeResponse{
-		Access: value.AllowedPledge{
+	return c.JSON(http.StatusOK, jsonmodels.AllowedManaPledgeResponse{
+		Access: jsonmodels.AllowedPledge{
 			IsFilterEnabled: access.IsFilterEnabled,
 			Allowed:         accessNodes,
 		},
-		Consensus: value.AllowedPledge{
+		Consensus: jsonmodels.AllowedPledge{
 			IsFilterEnabled: consensus.IsFilterEnabled,
 			Allowed:         consensusNodes,
 		},
