+++ plugins/webapi/mana/plugin.go
@@ -34,6 +34,7 @@ func configure(_ *node.Plugin) {
 	webapi.Server().GET("/mana/access/online", getOnlineAccessHandler)
 	webapi.Server().GET("/mana/consensus/online", getOnlineConsensusHandler)
 	webapi.Server().GET("/mana/pending", GetPendingMana)
+	webapi.Server().GET("mana/allowedManaPledge", allowedManaPledgeHandler)
 	// webapi.Server().GET("/mana/consensus/past", getPastConsensusManaVectorHandler)
 	// webapi.Server().GET("/mana/consensus/logs", getEventLogsHandler)
 	// webapi.Server().GET("/mana/consensus/metadata", getPastConsensusVectorMetadataHandler)
